refactor main router to app
