+++ src/App/App.tsx
@@ -1,12 +1,16 @@
-import './App.scss'
-import React, { useEffect } from 'react'
+import React, { useState, useEffect } from 'react'
 import MainNav from '../MainNav/MainNav'
 import ParkContainer from '../ParkContainer/ParkContainer'
+import SavedParks from '../SavedParks/SavedParks'
 import Footer from '../Footer/Footer'
 import { Route, Switch, useLocation, useHistory } from 'react-router-dom'
 import { nationalParks } from '../ParkData'
+import './App.scss'
+import { LocalParkContainer } from '../interfaces'
 
-const App: React.FC = () => {
+const App: React.FC<LocalParkContainer> = () => {
+  const [visitedList, setVisitedList] = useState<LocalParkContainer[] | any>([])
+  const [bucketList, setBucketList] = useState<LocalParkContainer[] | any>([])
   const location = useLocation()
   const history = useHistory()
 
@@ -16,9 +20,72 @@ const App: React.FC = () => {
     } 
   })
 
+  useEffect(() => {
+    retrieveFromStorage()
+  }, [])
+
+  useEffect(() => {
+    saveToStorage()
+  }, [visitedList, bucketList])
+
   const generateRandomParkCode = (): void => {
     let index = Math.floor(Math.random() * nationalParks.length)
     history.push(`/park/${nationalParks[index].parkCode}`)
+  }
+
+  const findChosenPark = (parkCode: string): any | void => {
+    const chosenPark = nationalParks.find(park => {
+      return park.parkCode === parkCode
+    })
+    if (chosenPark) {
+      return chosenPark
+    }
+  }
+
+  const addToVisited = (parkCode: string) => {
+    const addedPark = findChosenPark(parkCode)
+    addedPark.visited = true
+    if (!visitedList.includes(addedPark)) {
+      setVisitedList([...visitedList, addedPark])
+    }
+  }
+
+  const deleteFromVisited = (parkCode: string) => {
+    const removedPark = findChosenPark(parkCode)
+    removedPark.visited = false
+    const updatedParks = visitedList.filter((park: any) => park.parkCode !== parkCode)
+    setVisitedList(updatedParks)
+  }
+
+  const addToBucketList = (parkCode: string) => {
+    const addedPark = findChosenPark(parkCode)
+    addedPark.bucketList = true
+    if (!bucketList.includes(addedPark)) {
+      setBucketList([...bucketList, addedPark])
+    }
+  }
+
+  const deleteFromBucketList = (parkCode: string) => {
+    const removedPark = findChosenPark(parkCode)
+    removedPark.bucketList = false
+    const updatedParks = bucketList.filter((park: any) => park.parkCode !== parkCode)
+    setBucketList(updatedParks)
+  }
+
+  const saveToStorage = () => {
+    localStorage.clear()
+    let stringifiedVisited = JSON.stringify(visitedList)
+    let stringifiedBucketList = JSON.stringify(bucketList)
+    localStorage.setItem(`visitedList`, stringifiedVisited)
+    localStorage.setItem(`bucketList`, stringifiedBucketList)
+  }
+
+  const retrieveFromStorage = () => {
+    const storedVisited: any = localStorage.getItem('visitedList')
+    const parsedVisited = JSON.parse(storedVisited)
+    console.log(parsedVisited)
+    setVisitedList(parsedVisited)
+    saveToStorage()
   }
 
   return (
@@ -30,21 +97,32 @@ const App: React.FC = () => {
           <Route
             path='/park/:parkCode'
             render={({ match }) => {
-              const chosenPark = nationalParks.find(park => {
-                return park.parkCode === match.params.parkCode
-              })
-              let parkCode = !chosenPark ? generateRandomParkCode() : match.params.parkCode
-              return (
+               return (
                 <ParkContainer
-                  parkCode={match.params.parkCode} />
+                  parkCode={match.params.parkCode}
+                  visitedList={visitedList}
+                  bucketList={bucketList}
+                  findChosenPark={findChosenPark}
+                  addToVisited={addToVisited}
+                  deleteFromVisited={deleteFromVisited}
+                  addToBucketList={addToBucketList}
+                  deleteFromBucketList={deleteFromBucketList}
+                />
               )
             }}
           />
           <Route
-            path='/user/:list'
+            path='/user/visited'
             render={() => {
               return (
-                <ParkContainer parkCode={''}/>
+                <SavedParks
+                  visitedList={visitedList}
+                  bucketList={bucketList}
+                  addToVisited={addToVisited}
+                  deleteFromVisited={deleteFromVisited}
+                  addToBucketList={addToBucketList}
+                  deleteFromBucketList={deleteFromBucketList}
+                />
               )
             }}
           />
