+++ src/UserNav/UserNav.tsx
@@ -6,13 +6,13 @@ const UserNav: React.FC<CurrentParkContainer> = props => {
   const { currentPark } = props as CurrentParkContainer
   return (
     <nav className="user-nav">
-      <Link to={`/${currentPark.parkCode}/about`}>
+      <Link to={`/park/${currentPark.parkCode}/about`}>
         <h2>About</h2>
       </Link>
-      <Link to={`/${currentPark.parkCode}/location`}>
+      <Link to={`/park/${currentPark.parkCode}/location`}>
         <h2>Location</h2>
       </Link>
-      <Link to={`/${currentPark.parkCode}/contact`}>
+      <Link to={`/park/${currentPark.parkCode}/contact`}>
         <h2>Contact</h2>
       </Link>
     </nav>
